Move configs that are (sub)shell specific to bashrc
The profile file contained a number of definitions that I want to apply
to all shells (login and non-login), but by keeping them in profile,
they would only apply to login-shells. In particular, function
definitions (which includes bash completions) defined in profile would
not have effect in any non-login shells.

This commit fixes this by moving all of the affected configuration to
bashrc.