Home bin path and locale config for fish
Note the need for the '-P' option to fish_add_path. Without this, fish
would consider this a change to an universal variable and make changes
to fish/fish_variables. However, I want this command to keep working
even if the value of the HOME environment variable should ever change,
so I need -P to directly change the path instead.